Move server/config/secrets -> config/secrets
