+++ server/config/passport/google.js
@@ -1,5 +1,5 @@
 import { OAuth2Strategy as GoogleStrategy } from 'passport-google-oauth';
-import { google } from '../secrets';
+import { google } from '../../../config/secrets';
 import unsupportedMessage from '../../db/unsupportedMessage';
 import { passport as dbPassport } from '../../db';
